fix(fs): make fs.move use os.rename for files
This preserves permissions so should hopefully fix the issue with lpm
compile binaries not being executable.

+++ packages/fs/src/init.lua
@@ -100,6 +100,12 @@ end
 ---@param old string
 ---@param new string
 function fs.move(old, new)
+	-- Simpler and saves file permissions.
+	if fs.isfile(old) then
+		return os.rename(old, new)
+	end
+
+	-- Fallback to copy+delete for folders
 	if not fs.copy(old, new) then return false, "Failed to copy" end
 	if not fs.delete(old) then return false, "Failed to delete" end
 	return true
